pip_requirements.txt
Django==1.2
lxml
pycrypto
titlecase==0.5.1
